requirements.txt
openai
faiss-cpu
langchain
streamlit
pydantic
pyvis
neo4j
tqdm
python-dotenv
sentence-transformers
pytest
pyyaml
torch
pytest-mock
neo4j-driver